+++ geom/type_multi_line_string.go
@@ -510,3 +510,18 @@ func (m MultiLineString) Simplify(threshold float64) MultiLineString {
 	}
 	return NewMultiLineString(lss)
 }
+
+// SnapToGrid returns a copy of the MultiLineString with all coordinates
+// snapped to a base 10 grid.
+//
+// The grid spacing is specified by the number of decimal places to round to
+// (with negative decimal places being allowed). E.g., a decimalPlaces value of
+// 2 would cause all coordinates to be rounded to the nearest 0.01, and a
+// decimalPlaces of -1 would cause all coordinates to be rounded to the nearest
+// 10.
+//
+// Returned MultiLineStrings may be invalid due to snapping, even if the input
+// geometry was valid.
+func (m MultiLineString) SnapToGrid(decimalPlaces int) MultiLineString {
+	return m.TransformXY(snapToGridXY(decimalPlaces))
+}