+++ geom/type_polygon.go
@@ -681,3 +681,18 @@ func (p Polygon) Simplify(threshold float64, nv ...NoValidate) (Polygon, error) 
 	}
 	return simpl, nil
 }
+
+// SnapToGrid returns a copy of the Polygon with all coordinates snapped to a
+// base 10 grid.
+//
+// The grid spacing is specified by the number of decimal places to round to
+// (with negative decimal places being allowed). E.g., a decimalPlaces value of
+// 2 would cause all coordinates to be rounded to the nearest 0.01, and a
+// decimalPlaces of -1 would cause all coordinates to be rounded to the nearest
+// 10.
+//
+// Returned Polygons may be invalid due to snapping, even if the input geometry
+// was valid.
+func (p Polygon) SnapToGrid(decimalPlaces int) Polygon {
+	return p.TransformXY(snapToGridXY(decimalPlaces))
+}